Create tempdir in target basedir
If we create a `vendor/` directory in many git repositories, it
might not be `.gitignored` and hence may cause the IDE to try
to monitor all the files.

Instead, if one does `cargo vendor-filterer --format=tar.zstd target/vendor.tar.zstd`
as encourged by the docs, we now create the temporary directory in
`target` which will be gitignored *and* also automatically cleaned
up with `cargo clean`.