Add support for `tier` filtering
When we originally added support for platform globbing via e.g.
`*-unknown-linux-gnu` it worked with the Rust toolchain at the time.
Later, Rust added `m68k-unknown-linux-gnu` as a tier 3 target, which
is not installed by default.

This broke anyone using globs.

Add support for filtering by tiers; see
https://doc.rust-lang.org/nightly/rustc/platform-support.html

AFAICS, the list isn't available in any machine-readable form,
but it doesn't change often, so hardcode it for now.

This filtering combines with the platform one, which fixes the
bug.

Filtering by tier can also be used independently; this drops out
things like `redox_syscall` for example.

Closes: https://github.com/coreos/cargo-vendor-filterer/issues/50