Use `tar` and `flate2` crates
This gives us support for `--format=tar` and `--format=tar.gz`
on all platforms without an external binary.

For now, `zstd` compression requires an external binary because
the Rust crate still requires an external shared library, which
will be painful for Windows users.

I think this will be OK for MacOS users though since they're likely
to be using homebrew.

It may make sense to have a feature flag for use of external binaries.