Correctly handle filtered-out latest versions
Basically, in the case that there are two versions of a crate:

- foo-1
- foo-2

`cargo vendor` will write directories named:

- `vendor/foo-1`
- `vendor/foo` (meaning `foo-2`)

We weren't handling the case where the *newer* version of
these was filtered out.

To correctly handle this, we need to always gather an unfiltered
view to know what `cargo vendor` will do, and then build up
a mapping from package IDs to that name.

Closes: https://github.com/coreos/cargo-vendor-filterer/issues/75