fix(ci): nushell may want to default-no-clipboard and system-clipboard. enabling them as features.
we replace also `true` directory as `vendor`. I was confused of this at first... but it seems
`true` was used as to set all-features to true instead of it being an output directory

currently, to set `all-features` and `no-default-features`, you just have to add the flag
without any parameters. removing the flags will just auto set it to false.

Signed-off-by: Soc Virnyl Estela <contact@uncomfyhalomacro.pl>